% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{lts}
\alias{lts}
\alias{ltss}
\title{Syntactic Sugar for Tuning Space Construction}
\usage{
lts(.key, ...)

ltss(.keys, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}

\item{.keys}{(\code{character()})\cr
Keys passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve multiple objects.}
}
\value{
\itemize{
\item \link{TuningSpace} for \code{lts()}
\item list of \link{TuningSpace} for \code{ltss()}
}
}
\description{
This function complements \link{mlr_tuning_spaces} with functions in the spirit
of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
lts("classif.ranger.default")
}
