% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3tuningspaces-package}
\alias{mlr3tuningspaces}
\alias{mlr3tuningspaces-package}
\title{mlr3tuningspaces: Search Spaces for Hyperparameter Tuning}
\description{
Collection of search spaces for hyperparameter tuning. Includes various search spaces that can be directly applied on an `mlr3` learner. Additionally, meta information about the search space can be queried.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3tuningspaces.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3tuningspaces}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3tuningspaces/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Other contributors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID}) [contributor]
}

}
