% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuning_spaces.R
\docType{data}
\name{mlr_tuning_spaces}
\alias{mlr_tuning_spaces}
\title{Dictionary of Tuning Spaces}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{TuningSpace}.
Each tuning space has an associated help page, see \code{mlr_tuning_spaces_[id]}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link[mlr3misc:Dictionary]{mlr3misc::Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{\link[data.table:data.table]{data.table::data.table()}} with columns \code{"key"}, \code{"learner"} and
\code{"n_values"}.
}
}

\examples{
library(data.table)
as.data.table(mlr_tuning_spaces)
mlr_tuning_spaces$get("classif.ranger.default")
}
\concept{Dictionary}
\concept{TuningSpace}
\keyword{datasets}
