% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuning_spaces_classif.ranger.R
\name{mlr_tuning_spaces_classif.ranger}
\alias{mlr_tuning_spaces_classif.ranger}
\title{Tuning Space for the ranger Learner}
\source{
Bischl B, Binder M, Lang M, Pielok T, Richter J, Coors S, Thomas J, Ullmann T, Becker M, Boulesteix A (2021).
\emph{A Practical Introduction into Hyperparameter Optimization: Algorithms, Software, Best Practices and Open Challenges}.
Preprint.
}
\description{
A tuning space for \link[mlr3learners:mlr_learners_classif.ranger]{mlr3learners::LearnerClassifRanger}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.ranger}
\item Tuning Space:
\item replace [TRUE,FALSE]
\item sample.fraction \eqn{[0.1, 1]}{[0.1, 1]}
\item num.trees \eqn{[1, 2000]}{[1, 2000]}
}
}

