% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuning_spaces_classif.rpart.R
\name{mlr_tuning_spaces_classif.rpart}
\alias{mlr_tuning_spaces_classif.rpart}
\title{Tuning Space for the rpart Learner}
\source{
Bischl B, Binder M, Lang M, Pielok T, Richter J, Coors S, Thomas J, Ullmann T, Becker M, Boulesteix A (2021).
\emph{A Practical Introduction into Hyperparameter Optimization: Algorithms, Software, Best Practices and Open Challenges}.
Preprint.
}
\description{
A tuning space for \link[mlr3:mlr_learners_classif.rpart]{mlr3::LearnerClassifRpart}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.rpart}
\item Tuning Space:
\item minsplit (2, 128)
\item minbucket (1, 64)
\item cp (0.0001, 0.1)
}
}

