% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuning_spaces_classif.svm.R
\name{mlr_tuning_spaces_classif.svm}
\alias{mlr_tuning_spaces_classif.svm}
\title{Tuning Space for the svm Learner}
\source{
Bischl B, Binder M, Lang M, Pielok T, Richter J, Coors S, Thomas J, Ullmann T, Becker M, Boulesteix A (2021).
\emph{A Practical Introduction into Hyperparameter Optimization: Algorithms, Software, Best Practices and Open Challenges}.
Preprint.
}
\description{
A tuning space for \link[mlr3learners:mlr_learners_classif.svm]{mlr3learners::LearnerClassifSVM}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.svm}
\item Tuning Space:
\item cost \eqn{[1e-04, 10000]}{[1e-04, 10000]}
\item kernel [\dQuote{polynomial}, \dQuote{radial}, \dQuote{sigmoid}, \dQuote{linear}]
\item degree \eqn{[2, 5]}{[2, 5]}
\item gamma \eqn{[1e-04, 10000]}{[1e-04, 10000]}
}
}

