% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tuning_spaces_classif.xgboost.R
\name{mlr_tuning_spaces_classif.xgboost}
\alias{mlr_tuning_spaces_classif.xgboost}
\title{Tuning Space for the xgboost Learner}
\source{
Bischl B, Binder M, Lang M, Pielok T, Richter J, Coors S, Thomas J, Ullmann T, Becker M, Boulesteix A (2021).
\emph{A Practical Introduction into Hyperparameter Optimization: Algorithms, Software, Best Practices and Open Challenges}.
Preprint.
}
\description{
A tuning space for \link[mlr3learners:mlr_learners_classif.xgboost]{mlr3learners::LearnerClassifXgboost}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.xgboost}
\item Tuning Space:
\item eta \eqn{[1e-04, 1]}{[1e-04, 1]}
\item nrounds \eqn{[1, 5000]}{[1, 5000]}
\item max_depth \eqn{[1, 20]}{[1, 20]}
\item colsample_bytree \eqn{[0.1, 1]}{[0.1, 1]}
\item colsample_bylevel \eqn{[0.1, 1]}{[0.1, 1]}
\item lambda \eqn{[0.1, 1]}{[0.1, 1]}
\item gamma \eqn{[1e-04, 1000]}{[1e-04, 1000]}
\item alpha \eqn{[1e-04, 1000]}{[1e-04, 1000]}
\item subsample \eqn{[0.1, 1]}{[0.1, 1]}
}
}

