% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_rpart.R
\name{mlr_tuning_spaces_classif_rpart}
\alias{mlr_tuning_spaces_classif_rpart}
\title{Tuning Space for the rpart Learner}
\description{
A tuning space for the \link[mlr3:mlr_learners_classif.rpart]{mlr3::LearnerClassifRpart}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.rpart}
\item Tuning Space:
\item minsplit (2, 128)
\item minbucket (1, 64)
\item cp (0.0001, 0.1)
}
}

