% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_svm.R
\name{mlr_tuning_spaces_classif_svm}
\alias{mlr_tuning_spaces_classif_svm}
\title{Tuning Space for the svm Learner}
\description{
A tuning space for the \link[mlr3learners:mlr_learners_classif.svm]{mlr3learners::LearnerClassifSVM}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.svm}
\item Tuning Space:
\item cost (0.0001, 1e+04)
\item kernel (polynomial,radial,sigmoid\link{...})
\item degree (2, 5)
\item gamma (0.0001, 1e+04)
}
}

