% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_xgboost.R
\name{mlr_tuning_spaces_classif_xgboost}
\alias{mlr_tuning_spaces_classif_xgboost}
\title{Tuning Space for the xgboost Learner}
\description{
A tuning space for the \link[mlr3learners:mlr_learners_classif.xgboost]{mlr3learners::LearnerClassifXgboost}.
}
\section{Meta Information}{

\itemize{
\item Learner: \dQuote{classif.xgboost}
\item Tuning Space:
\item eta (0.0001, 1)
\item nrounds (1, 5000)
\item max_depth (1, 20)
\item colsample_bytree (0.1, 1)
\item colsample_bylevel (0.1, 1)
\item lambda (0.1, 1)
\item gamma (0.0001, 1000)
\item alpha (0.0001, 1000)
\item subsample (0.1, 1)
}
}

