% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_dl.R
\name{mlr_tuning_spaces_dl}
\alias{mlr_tuning_spaces_dl}
\alias{mlr_tuning_spaces_classif.ft_transformer.default}
\title{Deep Learning Tuning Spaces}
\source{
\verb{r format_bib("gorishniy2021revisting")}
}
\description{
Tuning spaces for deep neural network architectures from the \verb{r cite_bib("gorishniy2021revisting")} article.
}
