% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_default.R
\name{tuning_spaces_default}
\alias{tuning_spaces_default}
\alias{classif.glmnet.default}
\alias{classif.kknn.default}
\alias{classif.ranger.default}
\alias{classif.rpart.default}
\alias{classif.svm.default}
\alias{classif.xgboost.default}
\alias{regr.glmnet.default}
\alias{regr.kknn.default}
\alias{regr.ranger.default}
\alias{regr.rpart.default}
\alias{regr.svm.default}
\alias{regr.xgboost.default}
\title{Default Tuning Spaces}
\source{
Bischl B, Binder M, Lang M, Pielok T, Richter J, Coors S, Thomas J, Ullmann T, Becker M, Boulesteix A, Deng D, Lindauer M (2021).
\dQuote{Hyperparameter Optimization: Foundations, Algorithms, Best Practices and Open Challenges.}
2107.05847, \url{https://arxiv.org/abs/2107.05847}.
}
\description{
Tuning spaces from the Bischl (2021) article.
}
\section{glmnet tuning space}{

\itemize{
\item s \eqn{[1e-04, 10000]}{[1e-04, 10000]}
\item alpha \eqn{[0, 1]}{[0, 1]}
}
}

\section{kknn tuning space}{

\itemize{
\item k \eqn{[1, 50]}{[1, 50]}
\item distance \eqn{[1, 5]}{[1, 5]}
\item kernel [\dQuote{rectangular}, \dQuote{optimal}, \dQuote{epanechnikov}, \dQuote{biweight}, \dQuote{triweight}, \dQuote{cos}, \dQuote{inv}, \dQuote{gaussian}, \dQuote{rank}]
}
}

\section{ranger tuning space}{

\itemize{
\item mtry.ratio \eqn{[0, 1]}{[0, 1]}
\item replace [TRUE,FALSE]
\item sample.fraction \eqn{[0.1, 1]}{[0.1, 1]}
\item num.trees \eqn{[1, 2000]}{[1, 2000]}
}
}

\section{rpart tuning space}{

\itemize{
\item minsplit \eqn{[2, 128]}{[2, 128]}
\item minbucket \eqn{[1, 64]}{[1, 64]}
\item cp \eqn{[1e-04, 0.1]}{[1e-04, 0.1]}
}
}

\section{svm tuning space}{

\itemize{
\item cost \eqn{[1e-04, 10000]}{[1e-04, 10000]}
\item kernel [\dQuote{polynomial}, \dQuote{radial}, \dQuote{sigmoid}, \dQuote{linear}]
\item degree \eqn{[2, 5]}{[2, 5]}
\item gamma \eqn{[1e-04, 10000]}{[1e-04, 10000]}
}
}

\section{xgboost tuning space}{

\itemize{
\item eta \eqn{[1e-04, 1]}{[1e-04, 1]}
\item nrounds \eqn{[1, 5000]}{[1, 5000]}
\item max_depth \eqn{[1, 20]}{[1, 20]}
\item colsample_bytree \eqn{[0.1, 1]}{[0.1, 1]}
\item colsample_bylevel \eqn{[0.1, 1]}{[0.1, 1]}
\item lambda \eqn{[0.001, 1000]}{[0.001, 1000]}
\item alpha \eqn{[0.001, 1000]}{[0.001, 1000]}
\item subsample \eqn{[0.1, 1]}{[0.1, 1]}
}
}

