% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_package.R
\name{tuning_spaces_package}
\alias{tuning_spaces_package}
\alias{classif.kknn.package}
\alias{classif.ranger.package}
\alias{classif.rpart.package}
\alias{classif.svm.package}
\alias{classif.xgboost.package}
\alias{regr.kknn.package}
\alias{regr.ranger.package}
\alias{regr.rpart.package}
\alias{regr.svm.package}
\alias{regr.xgboost.package}
\title{Package Tuning Space}
\description{
This tuning space contains only default parameter values from the
upstream learner packages.
}
\section{ranger tuning space}{

\itemize{
\item num.trees [500]
\item mtry [NULL]
\item min.node.size [NULL]
\item max.depth [NULL]
\item replace [TRUE]
\item sample.fraction [1]
\item splitrule ["gini"]
}
}

\section{rpart tuning space}{

\itemize{
\item minsplit [20]
\item minbucket [7]
\item cp [0.01]
\item maxcompete [4]
\item maxsurrogate [5]
\item usesurrogate [2]
\item xval [10]
\item surrogatestyle [0]
\item maxdepth [30]
}
}

\section{svm tuning space}{

\itemize{
\item scale [TRUE]
\item type ["C-classification"]
\item kernel ["radial"]
\item cost [1]
\item tolerance [0.001]
\item epsilon [0.1]
\item shrinking [TRUE]
}
}

\section{xgboost tuning space}{

\itemize{
\item booster ["gbtree"]
\item eta [0.3]
\item gamma [0]
\item max_depth [6]
\item min_child_weight [1]
\item max_delta_step [0]
\item subsample [1]
\item colsample_bytree [1]
\item colsample_bylevel [1]
\item lambda [1]
\item alpha [0]
\item tree_method ["auto"]
\item scale_pos_weight [1]
\item refresh_leaf [TRUE]
\item process_type ["default"]
\item num_parallel_tree [1]
}
}

