% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{lts}
\alias{lts}
\alias{lts.character}
\alias{lts.Learner}
\alias{ltss}
\title{Syntactic Sugar for Tuning Space Construction}
\usage{
lts(x)

\method{lts}{character}(x)

\method{lts}{Learner}(x)

ltss(x)
}
\arguments{
\item{x}{(\code{character()} | \link[mlr3:Learner]{mlr3::Learner})\cr
If \code{character}, key passed the dictionary to retrieve the tuning space.
If \link[mlr3:Learner]{mlr3::Learner}, default tuning space is added to the learner.}
}
\value{
\itemize{
\item \link{TuningSpace} for \code{lts()}
\item list of \link{TuningSpace} for \code{ltss()}
}
}
\description{
This function complements \link{mlr_tuning_spaces} with functions in the spirit
of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
lts("classif.ranger.default")
}
