% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_sbb.R
\name{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_surv.akritas.sbb}
\alias{mlr_tuning_spaces_surv.glmnet.sbb}
\alias{mlr_tuning_spaces_surv.penalized.sbb}
\alias{mlr_tuning_spaces_surv.parametric.sbb}
\alias{mlr_tuning_spaces_surv.flexsurv.sbb}
\alias{mlr_tuning_spaces_surv.rfsrc.sbb}
\alias{mlr_tuning_spaces_surv.ranger.sbb}
\alias{mlr_tuning_spaces_surv.cforest.sbb}
\alias{mlr_tuning_spaces_surv.aorsf.sbb}
\alias{mlr_tuning_spaces_surv.rpart.sbb}
\alias{mlr_tuning_spaces_surv.mboost.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.cox.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.aft.sbb}
\alias{mlr_tuning_spaces_surv.svm.sbb}
\title{Burk (2024) Survival Tuning Spaces}
\source{
Burk L, Zobolas J, Bischl B, Bender A, Wright MN, Sonabend R (2024).
\dQuote{A Large-Scale Neutral Comparison Study of Survival Models on Low-Dimensional Data.}
2406.04098, \url{https://arxiv.org/abs/2406.04098}.
}
\description{
Tuning spaces from the Burk (2024) article.
}
\section{akritas tuning space}{

\itemize{
\item lambda \eqn{[0, 1]}{[0, 1]}
}
}

\section{cv_glmnet tuning space}{

\itemize{
\item alpha \eqn{[0, 1]}{[0, 1]}
}
}

\section{penalized tuning space}{

\itemize{
\item lambda1 \eqn{[-10, 10]}{[-10, 10]} (2^x)
\item lambda2 \eqn{[-10, 10]}{[-10, 10]} (2^x)
}
}

\section{parametric tuning space}{

\itemize{
\item dist [\dQuote{weibull}, \dQuote{exponential}, \dQuote{lognormal}, \dQuote{loglogistic}]
}
}

\section{flexsurv tuning space}{

\itemize{
\item k \eqn{[1, 10]}{[1, 10]}
}
}

\section{rfsrc tuning space}{

\itemize{
\item splitrule [\dQuote{bs.gradient}, \dQuote{logrank}]
\item mtry.ratio \eqn{[0, 1]}{[0, 1]}
\item nodesize \eqn{[1, 50]}{[1, 50]}
\item samptype [\dQuote{swr}, \dQuote{swor}]
\item sampsize.ratio \eqn{[0, 1]}{[0, 1]}
}
}

\section{ranger tuning space}{

\itemize{
\item splitrule [\dQuote{C}, \dQuote{maxstat}, \dQuote{logrank}]
\item mtry.ratio \eqn{[0, 1]}{[0, 1]}
\item min.node.size \eqn{[1, 50]}{[1, 50]}
\item replace [\if{html}{\out{<NULL>}}]
\item sample.fraction \eqn{[0, 1]}{[0, 1]}
}
}

\section{cforest tuning space}{

\itemize{
\item mtryratio \eqn{[0, 1]}{[0, 1]}
\item minsplit \eqn{[1, 50]}{[1, 50]}
\item mincriterion \eqn{[0, 1]}{[0, 1]}
\item replace [TRUE,FALSE]
\item fraction \eqn{[0, 1]}{[0, 1]}
}
}

\section{aorsf tuning space}{

\verb{r rd_info(lts("surv.aorsf.sbb"))}
}

\section{rpart tuning space}{

\itemize{
\item minbucket \eqn{[5, 50]}{[5, 50]}
}
}

\section{mboost tuning space}{

\itemize{
\item family [\dQuote{gehan}, \dQuote{cindex}, \dQuote{coxph}, \dQuote{weibull}]
\item mstop \eqn{[10, 5000]}{[10, 5000]}
\item nu \eqn{[0, 0.1]}{[0, 0.1]}
\item baselearner [\dQuote{bols}, \dQuote{btree}]
}
}

\section{xgboost.cox tuning space}{

\itemize{
\item max_depth \eqn{[1, 20]}{[1, 20]}
\item subsample \eqn{[0, 1]}{[0, 1]}
\item colsample_bytree \eqn{[0, 1]}{[0, 1]}
\item nrounds \eqn{[10, 5000]}{[10, 5000]}
\item eta \eqn{[0, 1]}{[0, 1]}
\item grow_policy [\dQuote{depthwise}, \dQuote{lossguide}]
}
}

\section{xgboost.aft tuning space}{

\itemize{
\item max_depth \eqn{[1, 20]}{[1, 20]}
\item subsample \eqn{[0, 1]}{[0, 1]}
\item colsample_bytree \eqn{[0, 1]}{[0, 1]}
\item nrounds \eqn{[10, 5000]}{[10, 5000]}
\item eta \eqn{[0, 1]}{[0, 1]}
\item grow_policy [\dQuote{depthwise}, \dQuote{lossguide}]
\item aft_loss_distribution [\dQuote{normal}, \dQuote{logistic}, \dQuote{extreme}]
\item aft_loss_distribution_scale \eqn{[0.5, 2]}{[0.5, 2]}
}
}

\section{ssvm tuning space}{

\verb{r rd_info(lts("surv.svm.sbb"))}
}

