% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_sbb.R
\name{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_surv.akritas.sbb}
\alias{mlr_tuning_spaces_surv.glmnet.sbb}
\alias{mlr_tuning_spaces_surv.penalized.sbb}
\alias{mlr_tuning_spaces_surv.parametric.sbb}
\alias{mlr_tuning_spaces_surv.flexsurv.sbb}
\alias{mlr_tuning_spaces_surv.rfsrc.sbb}
\alias{mlr_tuning_spaces_surv.ranger.sbb}
\alias{mlr_tuning_spaces_surv.cforest.sbb}
\alias{mlr_tuning_spaces_surv.aorsf.sbb}
\alias{mlr_tuning_spaces_surv.rpart.sbb}
\alias{mlr_tuning_spaces_surv.mboost.sbb}
\alias{mlr_tuning_spaces_surv.cv_coxboost.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.cox.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.aft.sbb}
\alias{mlr_tuning_spaces_surv.svm.sbb}
\title{Burk (2024) Survival Tuning Spaces}
\source{
Burk L, Zobolas J, Bischl B, Bender A, Wright MN, Sonabend R (2024).
\dQuote{A Large-Scale Neutral Comparison Study of Survival Models on Low-Dimensional Data.}
2406.04098, \url{https://arxiv.org/abs/2406.04098}.
}
\description{
Tuning spaces from the Burk (2024) article.
}
\section{akritas tuning space}{

\itemize{
\item lambda \eqn{[0, 1]}{[0, 1]}
}
}

\section{cv_glmnet tuning space}{

\itemize{
\item alpha \eqn{[0, 1]}{[0, 1]}
}
}

\section{penalized tuning space}{

\itemize{
\item lambda1 \eqn{[-10, 10]}{[-10, 10]} (2^x)
\item lambda2 \eqn{[-10, 10]}{[-10, 10]} (2^x)
}
}

\section{parametric tuning space}{

\verb{r rd_info(lts("surv.parametric.sbb"))}
}

\section{flexsurv tuning space}{

\itemize{
\item k \eqn{[1, 10]}{[1, 10]}
}
}

\section{rfsrc tuning space}{

\verb{r rd_info(lts("surv.rfsrc.sbb"))}
}

\section{ranger tuning space}{

\verb{r rd_info(lts("surv.ranger.sbb"))}
}

\section{cforest tuning space}{

\verb{r rd_info(lts("surv.cforest.sbb"))}
}

\section{aorsf tuning space}{

\verb{r rd_info(lts("surv.aorsf.sbb"))}
}

\section{rpart tuning space}{

\itemize{
\item minbucket \eqn{[5, 50]}{[5, 50]}
}
}

\section{mboost tuning space}{

\itemize{
\item family [\dQuote{gehan}, \dQuote{cindex}, \dQuote{coxph}, \dQuote{weibull}]
\item mstop \eqn{[10, 5000]}{[10, 5000]}
\item nu \eqn{[0, 0.1]}{[0, 0.1]}
\item baselearner [\dQuote{bols}, \dQuote{btree}]
}
}

\section{cv_coxboost tuning space}{

\verb{r rd_info(lts("surv.cv_coxboost.sbb"))}
}

\section{xgboost.cox tuning space}{

\verb{r rd_info(lts("surv.xgboost.cox.sbb"))}
}

\section{xgboost.aft tuning space}{

\verb{r rd_info(lts("surv.xgboost.aft.sbb"))}
}

\section{ssvm tuning space}{

\verb{r rd_info(lts("surv.svm.sbb"))}
}

