% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_spaces_sbb.R
\name{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_sbb}
\alias{mlr_tuning_spaces_surv.akritas.sbb}
\alias{mlr_tuning_spaces_surv.glmnet.sbb}
\alias{mlr_tuning_spaces_surv.penalized.sbb}
\alias{mlr_tuning_spaces_surv.parametric.sbb}
\alias{mlr_tuning_spaces_surv.flexsurv.sbb}
\alias{mlr_tuning_spaces_surv.rfsrc.sbb}
\alias{mlr_tuning_spaces_surv.ranger.sbb}
\alias{mlr_tuning_spaces_surv.cforest.sbb}
\alias{mlr_tuning_spaces_surv.aorsf.sbb}
\alias{mlr_tuning_spaces_surv.rpart.sbb}
\alias{mlr_tuning_spaces_surv.mboost.sbb}
\alias{mlr_tuning_spaces_surv.cv_coxboost.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.cox.sbb}
\alias{mlr_tuning_spaces_surv.xgboost.aft.sbb}
\alias{mlr_tuning_spaces_surv.svm.sbb}
\title{Burk (2024) Survival Tuning Spaces}
\source{
Burk L, Zobolas J, Bischl B, Bender A, Wright MN, Sonabend R (2024).
\dQuote{A Large-Scale Neutral Comparison Study of Survival Models on Low-Dimensional Data.}
2406.04098, \url{https://arxiv.org/abs/2406.04098}.
}
\description{
Tuning spaces from the Burk (2024) article.
}
\section{akritas tuning space}{

\itemize{
\item lambda \eqn{[0, 1]}{[0, 1]}
}
}

\section{cv_glmnet tuning space}{

\itemize{
\item alpha \eqn{[0, 1]}{[0, 1]}
}
}

\section{penalized tuning space}{

\verb{r rd_info(lts("surv.penalized.sbb"))}
}

\section{parametric tuning space}{

\itemize{
\item dist [\dQuote{weibull}, \dQuote{exponential}, \dQuote{lognormal}, \dQuote{loglogistic}]
}
}

\section{flexsurv tuning space}{

}

\section{rfsrc tuning space}{

}

\section{ranger tuning space}{

}

\section{cforest tuning space}{

}

\section{aorsf tuning space}{

}

\section{rpart tuning space}{

}

\section{mboost tuning space}{

}

\section{cv_coxboost tuning space}{

}

\section{xgboost.cox tuning space}{

}

\section{xgboost.aft tuning space}{

}

\section{ssvm tuning space}{

}

